requirements.txt
keras==2.7.0
keras_tcn==3.4.1
matplotlib==3.5.1
metrics==0.3.3
numpy==1.19.4
pandas==1.3.4
scikit_learn==1.0.2
talos==1.0
tensorflow==2.7.0
